requirements.txt
rasa==2.3.0
Django==2.2
djangorestframework==3.8.0
pyenchant==3.2.0
requests==2.25.1
profanity==1.1
django-cors-headers==3.7.0
gunicorn==20.0.4
